+++ lamw_manager/core/installer/installer.sh
@@ -2,8 +2,8 @@
 #-------------------------------------------------------------------------------------------------#
 #Universidade federal de Mato Grosso (Alma Mater)
 #Course: Science Computer
-#Version: 0.5.6
-#Date: 06/04/2023
+#Version: 0.5.7
+#Date: 09/11/2023
 #Description: "installer.sh" is part of the core of LAMW Manager. Contains routines for installing LAMW development environment
 #-------------------------------------------------------------------------------------------------#
 
@@ -24,17 +24,39 @@ singleCoreWarning(){
 	fi
 }
 
+checkUnsupportedCmdlineTools(){
+	local ret=1
+	local unsupported_cmdlinetools_version=10
+	local cmdline_tools_prop_path="$CMD_SDK_TOOLS_DIR/latest/source.properties"
+	if [ -e $cmdline_tools_prop_path ]; then 
+		local cmdline_tools_query="$(grep Pkg.Revision= $cmdline_tools_prop_path | awk -F= ' { print $NF }' ) > $unsupported_cmdlinetools_version "
+		local cmdline_tools_query_result=$(echo "$cmdline_tools_query" | bc)
+		[ "$cmdline_tools_query_result" = "1" ] && ret=0
+	fi
+}
 checkOldCmdlineTools(){
 	local ret=1
 	local cmdline_tools_prop_path="$CMD_SDK_TOOLS_DIR/latest/source.properties"
 	if [ -e $cmdline_tools_prop_path ]; then 
 		local cmdline_tools_query="$CMD_SDK_TOOLS_VERSION_STR > $(grep Pkg.Revision= $cmdline_tools_prop_path | awk -F= ' { print $NF }')"
 		local cmdline_tools_query_result=$(echo "$cmdline_tools_query" | bc)
-		[ $cmdline_tools_query_result = 1 ] && ret=0
+		[ "$cmdline_tools_query_result" = "1" ] && ret=0
 	fi
 
 	return $ret
 } 
+
+resolvesCmdlineToolsConflicts(){
+	if [ -e "$CMD_SDK_TOOLS_DIR/latest/package.xml" ]; then 
+		rm -rf "$CMD_SDK_TOOLS_DIR/latest/package.xml"
+	fi
+
+	if  ( checkOldCmdlineTools || checkUnsupportedCmdlineTools ); then 
+		if [ -e "$CMD_SDK_TOOLS_DIR/latest" ]; then 
+			rm -rf "$CMD_SDK_TOOLS_DIR/latest"
+		fi
+	fi
+}
 
 #set old Gradle from $LAMW_INSTALL_LOG
 setOldGradleVersion(){
@@ -103,10 +125,7 @@ LAMWPackageManager(){
 		fi
 	fi
 
-	if checkOldCmdlineTools ; then 
-		rm -rf "$CMD_SDK_TOOLS_DIR"
-	fi
-
+	resolvesCmdlineToolsConflicts
 
 }
 getStatusInstalation(){
@@ -240,12 +259,17 @@ GitClone(){
 
 
 GitPull(){
-	if [ "$git_branch" != "" ]; then 
-		gitCheckout
+	if [ "$git_branch" != "" ]; then
+		sucess_filler="$(getCurrentSucessFiller  4 $git_branch)"
+		startProgressBar 
+		if gitCheckout &>/dev/null; then
+			stopAsSuccessProgressBar
+		else stopProgressBarAsFail
+		fi
 	else 
 		changeDirectory "$git_src_dir"
 	fi
-	git config pull.ff only
+	git config pull.ff only	
 	git pull
 	GitReset
 }
@@ -365,7 +389,7 @@ getAndroidSDKTools(){
 	initROOT_LAMW
 	changeDirectory $ANDROID_SDK_ROOT
 	
-	if [ ! -e "$CMD_SDK_TOOLS_DIR" ];then
+	if [ ! -e "$CMD_SDK_TOOLS_DIR/latest" ];then
 		mkdir -p "$CMD_SDK_TOOLS_DIR"
 		changeDirectory "$CMD_SDK_TOOLS_DIR"
 		trap TrapControlC  2
@@ -438,8 +462,6 @@ getAndroidAPIS(){
 				runSDKManager ${SDK_MANAGER_CMD_PARAMETERS[i]} 
 			fi
 		done
-		
-		CmdLineToolsTrigger
 	else 
 		runSDKManager $*
 	fi
@@ -586,6 +608,7 @@ getMaxLAMWPackages(){
 	echo $max_lamw_pcks
 }
 
+
 installLAMWPackages(){
 	local ide_make_cfg_path="$LAMW_IDE_HOME_CFG/idemake.cfg"
 	local error_lazbuild_msg="${VERMELHO}Error${NORMAL}: Fails on build ${NEGRITO}${LAMW_PACKAGES[i]}${NORMAL} package"
@@ -602,27 +625,27 @@ installLAMWPackages(){
 		"--build-ide=" 
 		"--add-package"
 	)
+
+	local bg_pid=''
+
 	#build ide with lamw framework 
 	for((i=0;i< $max_lamw_pcks;i++)); do
 		
 		current_pack="`basename ${LAMW_PACKAGES[i]}`"
 		build_msg="Please wait, starting building ${NEGRITO}${current_pack}${NORMAL}..........................."
 		sucess_filler="$(getCurrentSucessFiller 3 $current_pack)"
-		
-		printf "%s" "$build_msg"
-		./lazbuild ${lamw_build_opts[*]} ${LAMW_PACKAGES[$i]} >/dev/null
-		
-		if [ $? != 0 ]; then 
-			printf  "%s\n" "${FILLER:${#sucess_filler}}${VERMELHO} [FAILS]${NORMAL}"
-			./lazbuild ${lamw_build_opts[*]} ${LAMW_PACKAGES[$i]}
-			[ $? != 0 ] && { echo "$error_lazbuild_msg" && EXIT_STATUS=1 && return ; }
+		startProgressBar
+		if ! ./lazbuild ${lamw_build_opts[*]} ${LAMW_PACKAGES[$i]} >/dev/null; then 
+			stopProgressBarAsFail
+			if ! ./lazbuild ${lamw_build_opts[*]} ${LAMW_PACKAGES[$i]}; then 
+			 	echo "$error_lazbuild_msg" 
+			 	EXIT_STATUS=1 
+			 	return 
+			fi
 		fi
 		
-		printf  "%s\n" "${FILLER:${#sucess_filler}}${VERDE} [OK]${NORMAL}"
-
+		stopAsSuccessProgressBar
 	done
-
-
 }
 #Build lazarus ide
 BuildLazarusIDE(){	
@@ -637,13 +660,14 @@ BuildLazarusIDE(){
 	changeDirectory $LAMW_IDE_HOME
 
 	if [ $# = 0 ]; then
-		printf "%s" "$build_msg"
+		startProgressBar
 		if ! make -s ${make_opts[@]} > /dev/null 2>&1; then
+			stopProgressBarAsFail
 			make -s ${make_opts[@]}
 			check_error_and_exit "$error_build_lazarus_msg" #build all IDE
 		fi
-	 	
-	 	printf  "%s\n" "${FILLER:${#sucess_filler}}${VERDE} [OK]${NORMAL}"
+	 	stopAsSuccessProgressBar
+	 	#printf  "%s\n" "${FILLER:${#sucess_filler}}${VERDE} [OK]${NORMAL}"
 	fi
 	
 	initLAMw4LinuxConfig
@@ -738,7 +762,7 @@ mainInstall(){
 	getGradle
 	getAndroidCmdLineTools
 	getFixLp
-	disableTrapActions
+	resetTrapActions
 	getFPCBuilder
 	getFPCSourcesTrunk
 	getLazarusSources